Many Changes

+++ sys/kern/types.h
@@ -12,5 +12,5 @@ typedef int32_t		s32;
 typedef int64_t		s64;
 
 typedef u32 blkno_t;
-typedef u16 devno_t;
+typedef u32 devno_t;
